requirements.txt
streamlit
easyocr
pillow
pandas
psycopg2-binary